requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
slack-sdk
langchain
langchain-google-vertexai
google-cloud-aiplatform
google-cloud-secret-manager
google-cloud-firestore
pandas
numpy
pydantic-settings
pytz
python-dateutil
youtube-transcript-api
yt-dlp
google-api-python-client
pytube
